+++ src/main/java/com/example/fcm/infra/redis/RedisTemplateManager.java
@@ -1,4 +1,4 @@
-package com.example.fcm.infra.fcm;
+package com.example.fcm.infra.redis;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
